api/requirements.txt
Flask==1.0.2
flask-cors
python-dotenv
psycopg2
